Fix xcode conflicts
