+++ CampusEventApp/CampusEventApp/MainView.swift
@@ -25,14 +25,13 @@ struct MainView: View {
                 
                 ChatView()
                     .tabItem {
-                    .tabItem {
                         Image(systemName: "message.fill")
                         Text("Chats")
                     }
                 
                 EventsView()
                     .tabItem {
-                        Image(systemName: "calendar")
+                        Image(systemName: "network")
                         Text("Events")
                     }
                 
@@ -66,30 +65,10 @@ struct MainView: View {
         let oneYearAgo = Date().addingTimeInterval(-365*24*60*60)
         let oneYearAfter = Date().addingTimeInterval(365*24*60*60)
         let predicate = eventStore.predicateForEvents(withStart: oneYearAgo, end: oneYearAfter, calendars: nil)
-
+        
         DispatchQueue.main.async {
             self.events = eventStore.events(matching: predicate)
             print("Fetched \(self.events.count) events")
-        }
-    }
-}
-
-struct CalendarTabView: View {
-    @Binding var eventStore: EKEventStore
-    @Binding var calendarAccessGranted: Bool
-    @Binding var events: [EKEvent]
-
-    var body: some View {
-        NavigationView {
-            List {
-                NavigationLink(destination: ImportCalendarView(eventStore: $eventStore, calendarAccessGranted: $calendarAccessGranted, events: $events)) {
-                    Label("Import Calendar", systemImage: "square.and.arrow.down")
-                }
-                NavigationLink(destination: InAppCalendarView(events: $events)) {
-                    Label("Calendar", systemImage: "calendar")
-                }
-            }
-            .navigationTitle("Calendar")
         }
     }
 }
