--- CampusEventApp/CampusEventApp/Views/NewsView.swift
@@ -1,27 +1,0 @@
-//
-//  NewsView.swift
-//  CampusEventApp
-//
-//  Created by Reinardus on 18.05.24.
-//
-
-import SwiftUI
-
-struct NewsView: View {
-    var body: some View {
-        ZStack {
-            Image("hintergrund")
-                .resizable()
-                .scaledToFill()
-                .ignoresSafeArea()
-            
-            Image(systemName: "bell.fill")
-                .foregroundColor(Color.black)
-                .font(.system(size: 100.0))
-        }
-    }
-}
-
-#Preview {
-    NewsView()
-}
